requirements.txt
python-chess==1.999.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
tqdm==4.65.0
PyYAML==6.0.1 